Move geometry functions into an isolated header
